+++ gnuradio-runtime/lib/vmcircbuf_mmap_shm_open.cc
@@ -30,7 +30,7 @@
 
 namespace gr {
 
-vmcircbuf_mmap_shm_open::vmcircbuf_mmap_shm_open(int size) : gr::vmcircbuf(size)
+vmcircbuf_mmap_shm_open::vmcircbuf_mmap_shm_open(size_t size) : gr::vmcircbuf(size)
 {
 #if !defined(HAVE_MMAP) || !defined(HAVE_SHM_OPEN)
     std::stringstream error_msg;
@@ -178,7 +178,7 @@ gr::vmcircbuf_factory* vmcircbuf_mmap_shm_open_factory::singleton()
 
 int vmcircbuf_mmap_shm_open_factory::granularity() { return gr::pagesize(); }
 
-gr::vmcircbuf* vmcircbuf_mmap_shm_open_factory::make(int size)
+gr::vmcircbuf* vmcircbuf_mmap_shm_open_factory::make(size_t size)
 {
     try {
         return new vmcircbuf_mmap_shm_open(size);
